+++ python/tests/test_hadamard_test.py
@@ -18,8 +18,6 @@ from qdk_chemistry.algorithms.hadamard_test.hadamard_test import HadamardTestBas
 from qdk_chemistry.data import AlgorithmRef, Circuit, MajoranaMapping, Structure, UnitaryRepresentation, Wavefunction
 from qdk_chemistry.utils.pauli_matrix import pauli_to_dense_matrix
 
-pytestmark = pytest.mark.usefixtures("use_base_qdk_ctx")
-
 _HAS_QSHARP = importlib.util.find_spec("qdk.qsharp") is not None
 
 _SHOTS = 5000